adjustmments and changes to file names for code cleannes and organization
