+++ unimarket/lib/Models/Repository/productReposirory.dart
@@ -1,9 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:unimarket/Model/model.dart';
-import 'package:unimarket/Model/product_model.dart';
-import 'package:unimarket/Model/seller_model.dart';
+import 'package:unimarket/Models/model.dart';
+import 'package:unimarket/Models/product_model.dart';
+import 'package:unimarket/Models/seller_model.dart';
 import 'package:uuid/uuid.dart';
 
 class ProductRepository {
