+++ unimarket/lib/Views/login_view.dart
@@ -1,67 +1,22 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:unimarket/Views/vista_login.dart';
-import 'package:unimarket/Controllers/auth.dart';
+import 'package:unimarket/Controllers/search_controller.dart';
+import 'package:unimarket/Views/home_view.dart';
+import 'package:unimarket/Views/register_view.dart';
+import 'package:unimarket/Controllers/auth_controller.dart';
+import 'package:unimarket/Views/body_view.dart';
 
-class VistaRegistrarse extends StatefulWidget {
-  const VistaRegistrarse({super.key});
+class LoginView extends StatefulWidget {
+  const LoginView({super.key});
+
   @override
-  EstadoRegistrarse createState() => EstadoRegistrarse();
+  EstadoLogin createState() => EstadoLogin();
 }
 
-class EstadoRegistrarse extends State<VistaRegistrarse> {
+class EstadoLogin extends State<LoginView> {
   String email = "";
   String contrasena = "";
-  AuthService autenticador = AuthService();
-  late bool registroExitoso;
-
-  void manejarRegistro() {}
-
-  void showErrorDialog(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          backgroundColor: Colors.red, // Set background color to red
-          title: Text('Oops!'),
-          content: Text(
-              'Something went wrong registering your user. Make sure you are using a valid email format. Make sure your password is at least 6 characters long. Make sure your email is not associated with an existing account'),
-          actions: <Widget>[
-            TextButton(
-              onPressed: () {
-                // Close the dialog
-                Navigator.of(context).pop();
-              },
-              child: Text('OK'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void showSuccessDialog(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          backgroundColor: Colors.green, // Set background color to green
-          title: Text('Success!'),
-          content: Text('Registration successful! Now please sign in'),
-          actions: <Widget>[
-            TextButton(
-              onPressed: () {
-                // Close the dialog
-                Navigator.push(context,
-                    MaterialPageRoute(builder: (context) => VistaLogin()));
-              },
-              child: Text('OK'),
-            ),
-          ],
-        );
-      },
-    );
-  }
+  AuthController autenticador = AuthController();
 
 //hace parte del scaffold. ignorar para otras cosas
   Widget _buildTextField(String labelText, String hintText,
@@ -138,22 +93,45 @@ class EstadoRegistrarse extends State<VistaRegistrarse> {
     );
   }
 
+  void showErrorDialog(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          backgroundColor: Colors.red, // Set background color to red
+          title: Text('Oops!'),
+          content: Text(
+              'Something went wrong singing in. Make sure the credentials you are providing are correct and that you already have an account registered'),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () {
+                // Close the dialog
+                Navigator.of(context).pop();
+              },
+              child: Text('OK'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
+    void authenticationProcess(bool existing_user) {
+      // Aquí va el proceso de verificación con Firebase
+
+      if (existing_user) {
+        SearchControllerUnimarket().cargarProductos();
+        Navigator.push(
+            context, MaterialPageRoute(builder: (context) => BodyView()));
+      } else {
+        showErrorDialog(context);
+      }
+    }
+
     return Scaffold(
       backgroundColor: const Color.fromARGB(255, 250, 206, 190),
-      appBar: AppBar(
-        leading: IconButton(
-          icon: Icon(Icons.arrow_back),
-          onPressed: () {
-            Navigator.push(
-                context, MaterialPageRoute(builder: (context) => VistaLogin()));
-          },
-        ),
-        backgroundColor:
-            Colors.transparent, // Para hacer transparente la AppBar
-        elevation: 0, // Para eliminar la sombra debajo de la AppBar
-      ),
       body: Column(
         children: [
           Container(
@@ -172,7 +150,8 @@ class EstadoRegistrarse extends State<VistaRegistrarse> {
             child: Center(
               child: Container(
                 width: 350,
-                margin: const EdgeInsets.only(bottom: 100),
+                height: 550,
+                // margin: const EdgeInsets.only(bottom: 100),
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(20),
                   color: const Color.fromARGB(255, 255, 100, 35),
@@ -186,7 +165,7 @@ class EstadoRegistrarse extends State<VistaRegistrarse> {
                         const Padding(
                           padding: EdgeInsets.only(bottom: 35),
                           child: Text(
-                            'Register',
+                            'Sign in',
                             style: TextStyle(
                               fontSize: 36.0,
                               fontWeight: FontWeight.bold,
@@ -201,23 +180,39 @@ class EstadoRegistrarse extends State<VistaRegistrarse> {
                         const SizedBox(height: 20.0),
                         ElevatedButton(
                           onPressed: () async {
-                            // try {
-                            autenticador.registrar(email, contrasena).then(
-                                (value) => value == null
-                                    ? showErrorDialog(context)
-                                    : showSuccessDialog(context));
-                            // } catch (e) {
-                            //   print(e.toString() + "9999999999999999999999");
-                            //   showErrorDialog(context);
-                            // }
-
-                            // print(registroExitoso);
-                            // registroExitoso
-                            //     ? showErrorDialog(context)
-                            // showErrorDialog(context);
+                            authenticationProcess(
+                                await autenticador.ingresar(email, contrasena));
                           },
                           style: ElevatedButton.styleFrom(
                             backgroundColor: Colors.white,
+                            padding:
+                                const EdgeInsets.symmetric(horizontal: 40.0),
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(10.0),
+                            ),
+                          ),
+                          child: const Text(
+                            'Sign in',
+                            style: TextStyle(
+                              fontSize: 15,
+                              fontWeight: FontWeight.bold,
+                              color: Colors.black,
+                            ),
+                          ),
+                        ),
+                        const SizedBox(
+                            height:
+                                20.0), // Add some space between the Sign in button and Register button
+                        ElevatedButton(
+                          onPressed: () {
+                            // Add your registration logic here
+                            Navigator.push(
+                                context,
+                                MaterialPageRoute(
+                                    builder: (context) => RegisterView()));
+                          },
+                          style: ElevatedButton.styleFrom(
+                            backgroundColor: Colors.orange,
                             padding:
                                 const EdgeInsets.symmetric(horizontal: 40.0),
                             shape: RoundedRectangleBorder(
