+++ unimarket/lib/main.dart
@@ -3,10 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:unimarket/Views/body_view.dart';
 import 'package:unimarket/Views/home_view.dart';
 import 'package:unimarket/Views/publish_view.dart';
-import 'package:unimarket/Views/vista_login.dart';
+import 'package:unimarket/Views/login_view.dart';
 import 'package:unimarket/firebase_options.dart';
-import 'package:unimarket/Views/vista_login.dart';
-import 'package:unimarket/Views/vista_registrarse.dart';
+import 'package:unimarket/Views/login_view.dart';
+import 'package:unimarket/Views/register_view.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/services.dart';
 
@@ -33,9 +33,8 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
-        
       ),
-      home: VistaLogin(),
+      home: LoginView(),
       //home: HomeView(),
       //home: PublishView(),
     );
